chapter03 updated
