+++ chapter03/dags/12_no_catchup.py
@@ -2,12 +2,15 @@ from datetime import datetime
 from pathlib import Path
 
 import pandas as pd
+
 from airflow import DAG
 from airflow.operators.bash import BashOperator
 from airflow.operators.python import PythonOperator
 
-def _calculate_stats(input_path, output_path):
+def _calculate_stats(**context):
     """Calculates event statistics."""
+    input_path = context["templates_dict"]["input_path"]
+    output_path = context["templates_dict"]["output_path"]
 
     events = pd.read_json(input_path)
     stats = events.groupby(["date", "user"]).size().reset_index()
@@ -16,29 +19,33 @@ def _calculate_stats(input_path, output_path):
     stats.to_csv(output_path, index=False)
 
 
-
 with DAG(
-    dag_id="05_query_with_dates",
+    dag_id="12_no_catchup",
     schedule="@daily",
-    start_date=datetime(year=2019, month=1, day=1),
-    end_date=datetime(year=2019, month=1, day=5),
+    start_date=datetime(2019,1,1),
+    end_date=datetime(2019,1,5),
+    catchup=False,
 ):
 
     fetch_events = BashOperator(
         task_id="fetch_events",
         bash_command=(
             "mkdir -p /data/events && "
-            "curl -o /data/events.json "
+            "curl -o /data/events/{{ds}}.json "
             "http://events_api:5000/events?"
-            "start_date=2019-01-01&"
-            "end_date=2019-01-02"
+            "start_date={{ds}}&"
+            "end_date={{next_ds}}"
         ),
     )
 
     calculate_stats = PythonOperator(
         task_id="calculate_stats",
         python_callable=_calculate_stats,
-        op_kwargs={"input_path": "/data/events.json", "output_path": "/data/stats.csv"},
+        templates_dict={
+            "input_path": "/data/events/{{ds}}.json",
+            "output_path": "/data/stats/{{ds}}.csv",
+        },
     )
+
 
     fetch_events >> calculate_stats
