+++ chapter03/dags/13_non_atomic_send.py
@@ -7,8 +7,11 @@ from airflow import DAG
 from airflow.operators.bash import BashOperator
 from airflow.operators.python import PythonOperator
 
-def _calculate_stats(input_path, output_path):
+
+def _calculate_stats(**context):
     """Calculates event statistics."""
+    input_path = context["templates_dict"]["input_path"]
+    output_path = context["templates_dict"]["output_path"]
 
     events = pd.read_json(input_path)
     stats = events.groupby(["date", "user"]).size().reset_index()
@@ -16,30 +19,41 @@ def _calculate_stats(input_path, output_path):
     Path(output_path).parent.mkdir(exist_ok=True)
     stats.to_csv(output_path, index=False)
 
+    _email_stats(stats, email="user@example.com")
+
+
+def _email_stats(stats, email):
+    """Send an email..."""
+    print(f"Sending stats to {email}...")
+
 
 
 with DAG(
-    dag_id="06_templated_query",
-    schedule_interval="@daily",
+    dag_id="13_non_atomic_send",
+    schedule="@daily",
     start_date=datetime(year=2019, month=1, day=1),
     end_date=datetime(year=2019, month=1, day=5),
+    catchup=True,
 ):
 
     fetch_events = BashOperator(
         task_id="fetch_events",
         bash_command=(
             "mkdir -p /data/events && "
-            "curl -o /data/events.json "
+            "curl -o /data/events/{{ds}}.json "
             "http://events_api:5000/events?"
-            "start_date={{execution_date.strftime('%Y-%m-%d')}}&"
-            "end_date={{next_execution_date.strftime('%Y-%m-%d')}}"
+            "start_date={{ds}}&"
+            "end_date={{next_ds}}"
         ),
     )
 
     calculate_stats = PythonOperator(
         task_id="calculate_stats",
         python_callable=_calculate_stats,
-        op_kwargs={"input_path": "/data/events.json", "output_path": "/data/stats.csv"},
+        templates_dict={
+            "input_path": "/data/events/{{ds}}.json",
+            "output_path": "/data/stats/{{ds}}.csv",
+        },
     )
 
     fetch_events >> calculate_stats
